first set of working tests
