--- gbe/fixtures/special_privs.json
@@ -1,122 +1,0 @@
-[
-{
-    "pk": 1, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Act Coordinator", 
-        "permissions": []
-    }
-},
-{
-    "pk": 2, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Act Reviewers", 
-        "permissions": []
-    }
-},
-{
-    "pk": 3, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Class Coordinator", 
-        "permissions": []
-    }
-},
-{
-    "pk": 4, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Class Reviewers", 
-        "permissions": []
-    }
-},
-{
-    "pk": 5, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Vendor Coordinator", 
-        "permissions": []
-    }
-},
-{
-    "pk": 6, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Vendor Reviewers", 
-        "permissions": []
-    }
-},
-{
-    "pk": 7, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Volunteer Coordinator", 
-        "permissions": []
-    }
-},
-{
-    "pk": 8, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Volunteer Reviewers", 
-        "permissions": []
-    }
-},
-{
-    "pk": 9, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Registrar", 
-        "permissions": []
-    }
-},
-{
-    "pk": 10, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Scheduling Mavens", 
-        "permissions": []
-    }
-},
-{
-    "pk": 11, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Ticketing - Admin", 
-        "permissions": []
-    }
-},
-{
-    "pk": 12, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Ticketing - Transactions", 
-        "permissions": []
-    }
-},
-{
-    "pk": 13, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Tech Crew", 
-        "permissions": []
-    }
-},
-{
-    "pk": 14, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Costume Coordinator", 
-        "permissions": []
-    }
-},
-{
-    "pk": 15, 
-    "model": "auth.group", 
-    "fields": {
-        "name": "Costume Reviewers", 
-        "permissions": []
-    }
-},
-]
