+++ gbe/views/review_flex_bid_view.py
@@ -37,7 +37,7 @@ from scheduler.idd import get_schedule
 class FlexibleReviewBidView(ReviewBidView):
     reviewer_permissions = ('Act Reviewers', )
     review_list_view_name = 'act_review_list'
-    coordinator_permissions = ('Act Coordinator',)
+    coordinator_permissions = ('Act Coordinator', 'Producer')
     bid_state_change_form = BidStateChangeForm
     bid_evaluation_type = FlexibleEvaluation
     bid_evaluation_form_type = FlexibleEvaluationForm
